requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
rdkit-pypi>=2023.9.1
Pillow>=10.0.0
